feat: move review stories under demos
