+++ packages/lib/src/components/Card/stories/Card.stories.tsx
@@ -10,7 +10,6 @@ import { CardWith3DS2CreateFromAction } from './cardStoryHelpers/CardWith3DS2Cre
 import { AdditionalDetailsActions, AdditionalDetailsData } from '../../../core/types';
 import { displayResultMessage } from '../../../../storybook/helpers/checkout-handlers';
 import { makeDetailsCall } from '../../../../storybook/helpers/checkout-api-calls';
-import { renderCardReviewPage } from '../../../../storybook/helpers/review-page-renders';
 
 type CardStory = StoryConfiguration<CardConfiguration>;
 
@@ -332,12 +331,6 @@ export const SplitFundingSourceTest: CardStory = {
             }
         }
     }
-};
-
-export const ReviewPage: CardStory = {
-    args: { countryCode: 'NL', useSessions: true },
-    tags: ['no-automated-visual-test'],
-    render: renderCardReviewPage
 };
 
 export default meta;
